# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * InfluenceHub API - Connect businesses with influencers, YouTubers, fashion models and celebrities
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TalentCategory =
  (typeof TalentCategory)[keyof typeof TalentCategory];

export const TalentCategory = {
  influencer: "influencer",
  youtuber: "youtuber",
  fashion_model: "fashion_model",
  celebrity: "celebrity",
  brand_ambassador: "brand_ambassador",
} as const;

export interface Talent {
  id: number;
  name: string;
  category: TalentCategory;
  bio: string;
  profileImageUrl?: string;
  coverImageUrl?: string;
  followerCount: number;
  engagementRate?: number;
  ratePerPost: number;
  ratePerCampaign?: number;
  platforms?: string[];
  niche?: string;
  location?: string;
  verified: boolean;
  available: boolean;
  managerName?: string;
  managerEmail?: string;
  avgRating?: number;
  totalBookings?: number;
  createdAt: string;
}

export interface CreateTalentBody {
  name: string;
  category: string;
  bio: string;
  profileImageUrl?: string;
  followerCount: number;
  engagementRate?: number;
  ratePerPost: number;
  ratePerCampaign?: number;
  platforms?: string[];
  niche?: string;
  location?: string;
  managerName?: string;
  managerEmail?: string;
}

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  pending: "pending",
  accepted: "accepted",
  rejected: "rejected",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type BookingCampaignType =
  (typeof BookingCampaignType)[keyof typeof BookingCampaignType];

export const BookingCampaignType = {
  social_post: "social_post",
  video_review: "video_review",
  brand_ambassador: "brand_ambassador",
  live_event: "live_event",
  product_launch: "product_launch",
} as const;

export interface Booking {
  id: number;
  talentId: number;
  talent?: Talent;
  businessName: string;
  businessEmail: string;
  campaignName: string;
  campaignDescription?: string;
  budget: number;
  status: BookingStatus;
  campaignType?: BookingCampaignType;
  startDate?: string;
  endDate?: string;
  createdAt: string;
}

export interface CreateBookingBody {
  talentId: number;
  businessName: string;
  businessEmail: string;
  campaignName: string;
  campaignDescription?: string;
  budget: number;
  campaignType: string;
  startDate?: string;
  endDate?: string;
}

export type UpdateBookingBodyStatus =
  (typeof UpdateBookingBodyStatus)[keyof typeof UpdateBookingBodyStatus];

export const UpdateBookingBodyStatus = {
  pending: "pending",
  accepted: "accepted",
  rejected: "rejected",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface UpdateBookingBody {
  status?: UpdateBookingBodyStatus;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  description?: string;
  iconName?: string;
  talentCount: number;
}

export type MessageSenderRole =
  (typeof MessageSenderRole)[keyof typeof MessageSenderRole];

export const MessageSenderRole = {
  business: "business",
  manager: "manager",
} as const;

export interface Message {
  id: number;
  bookingId: number;
  senderName: string;
  senderRole: MessageSenderRole;
  content: string;
  createdAt: string;
}

export interface CreateMessageBody {
  bookingId: number;
  senderName: string;
  senderRole: string;
  content: string;
}

export interface Review {
  id: number;
  talentId: number;
  bookingId?: number;
  reviewerName: string;
  reviewerCompany?: string;
  rating: number;
  comment?: string;
  createdAt: string;
}

export interface CreateReviewBody {
  talentId: number;
  bookingId?: number;
  reviewerName: string;
  reviewerCompany?: string;
  rating: number;
  comment?: string;
}

export interface PlatformStats {
  totalTalents: number;
  totalBookings: number;
  totalBusinesses: number;
  completedCampaigns: number;
  totalRevenue: number;
  activeBookings: number;
}

export interface CategoryStat {
  category: string;
  count: number;
  totalRevenue: number;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  booking_created: "booking_created",
  booking_accepted: "booking_accepted",
  booking_completed: "booking_completed",
  talent_joined: "talent_joined",
  review_posted: "review_posted",
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  description: string;
  entityName: string;
  createdAt: string;
}

export type ListTalentsParams = {
  category?: string;
  search?: string;
  minFollowers?: number;
  maxRate?: number;
};

export type ListBookingsParams = {
  status?: string;
};

export type ListMessagesParams = {
  bookingId: number;
};

export type ListReviewsParams = {
  talentId: number;
};