# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * InfluenceHub API - Connect businesses with influencers, YouTubers, fashion models and celebrities
 * OpenAPI spec version: 0.1.0
 */
import type { BookingCampaignType } from "./bookingCampaignType";
import type { BookingStatus } from "./bookingStatus";
import type { Talent } from "./talent";

export interface Booking {
  id: number;
  talentId: number;
  talent?: Talent;
  businessName: string;
  businessEmail: string;
  campaignName: string;
  campaignDescription?: string;
  budget: number;
  status: BookingStatus;
  campaignType?: BookingCampaignType;
  startDate?: string;
  endDate?: string;
  createdAt: string;
}