# lib/api-zod/src/generated/types/createBookingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * InfluenceHub API - Connect businesses with influencers, YouTubers, fashion models and celebrities
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBookingBody {
  talentId: number;
  businessName: string;
  businessEmail: string;
  campaignName: string;
  campaignDescription?: string;
  budget: number;
  campaignType: string;
  startDate?: string;
  endDate?: string;
}