# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * InfluenceHub API - Connect businesses with influencers, YouTubers, fashion models and celebrities
 * OpenAPI spec version: 0.1.0
 */

export interface Review {
  id: number;
  talentId: number;
  bookingId?: number;
  reviewerName: string;
  reviewerCompany?: string;
  rating: number;
  comment?: string;
  createdAt: string;
}